Lab01_02/script/requirements.txt
certifi==2019.11.28
chardet==3.0.4
gitdb==4.0.2
GitPython==3.1.0
idna==2.9
requests==2.23.0
smmap==3.0.1
urllib3==1.25.8
radon==4.1.0
